renamed surveyState to state
