+++ webapp/store/survey/nodeDefs/actions.js
@@ -19,7 +19,7 @@ import * as NodeDefState from '@webapp/loggedin/surveyViews/nodeDef/nodeDefState
 import { hideAppLoader, showAppLoader } from '@webapp/app/actions'
 import { showNotification } from '@webapp/app/appNotification/actions'
 import { showDialogConfirm } from '@webapp/app/appDialogConfirm/actions'
-import * as SurveyState from '../surveyState'
+import * as SurveyState from '../state'
 
 export const nodeDefCreate = 'survey/nodeDef/create'
 export const nodeDefUpdate = 'survey/nodeDef/update'
